Fix stability issues and improve reliability
- Add BLE queue overflow detection with throttled warnings
- Add strict bounds checking in alert parser to prevent buffer overflow
- Fix touch debounce millis() rollover at 49 days using signed comparison
- Only create BLE mutexes once to prevent resource leaks on restart
- Smart boot splash: show on cold boot, skip on crash reboots
- Improve BLE callback memory management (only allocate once)

+++ src/touch_handler.cpp
@@ -76,9 +76,9 @@ bool TouchHandler::isTouched() {
 }
 
 bool TouchHandler::getTouchPoint(int16_t& x, int16_t& y) {
-    // Check for debounce
+    // Check for debounce (handles millis() rollover at 49 days)
     unsigned long now = millis();
-    if (now - lastTouchTime < touchDebounceMs) {
+    if ((long)(now - lastTouchTime) < (long)touchDebounceMs) {
         return false;  // Still in debounce period
     }
